src/api/requirements.txt
fastapi==0.115.0
uvicorn==0.38.0

azure-identity==1.17.1
azure-ai-projects==2.0.0b2
azure-ai-agents==1.2.0b5

# Agent Framework preview bits require installation with --pre when using pip directly.
agent-framework-azure-ai==1.0.0b251120
